Refactored vulkan to decouple components that shouldn't be coupled.
