+++ nc/source/graphics/Graphics2.cpp
@@ -1,24 +1,20 @@
 #include "Graphics2.h"
 #include "vulkan/Base.h"
-#include "vulkan/Instance.h"
 #include "vulkan/GraphicsPipeline.h"
-#include "vulkan/RenderPass.h"
-#include "vulkan/FrameBuffers.h"
 #include "vulkan/Commands.h"
 #include "vulkan/VertexBuffer.h"
 #include "vulkan/IndexBuffer.h"
+#include "vulkan/Swapchain.h"
 
 namespace nc::graphics
 {
     using namespace vulkan;
 
     Graphics2::Graphics2(HWND hwnd, HINSTANCE hinstance, Vector2 dimensions)
-        : m_instance{ std::make_unique<Instance>(hwnd, hinstance) },
-          m_base{ std::make_unique<Base>(*m_instance, dimensions) },
-          m_renderPass{ std::make_unique<RenderPass>(*m_base) },
-          m_pipeline{ std::make_unique<GraphicsPipeline>(*m_base, *m_renderPass) },
-          m_frameBuffers{ std::make_unique<FrameBuffers>(*m_base, *m_renderPass) },
-          m_commands{ nullptr },
+        : m_base{ std::make_unique<Base>(hwnd, hinstance) },
+          m_swapchain{ std::make_unique<Swapchain>(m_base.get(), dimensions) },
+          m_pipeline{ std::make_unique<GraphicsPipeline>(*m_base, *m_swapchain) },
+          m_commands{ std::make_unique<Commands>(*m_base, *m_swapchain) },
           m_vertexBuffer{ nullptr }, // @todo: Take from mesh, will not be created in CTOR for Graphics2
           m_indexBuffer{ nullptr }, // @todo: Take from mesh, will not be created in CTOR for Graphics2
           m_dimensions{ dimensions },
@@ -27,16 +23,10 @@ namespace nc::graphics
           m_viewMatrix{},
           m_projectionMatrix{}
     {  
-        m_commands = std::make_unique<Commands>(*m_base,
-            m_base->GetSemaphores(SemaphoreType::RenderReady),
-            m_base->GetSemaphores(SemaphoreType::PresentReady),
-            m_base->GetFences(FenceType::FramesInFlight),
-            m_base->GetFences(FenceType::ImagesInFlight));
-
         // @todo: Take from mesh, will not be created in CTOR for Graphics2
         auto vertices = std::vector<vulkan::vertex::Vertex> 
         {
-            // Position            Color
+            // Position              Color
             { Vector2{-0.5f, -0.5f}, Vector3{1.0f, 1.0f, 1.0f} },
             { Vector2{0.5f, -0.5f},  Vector3{0.0f, 0.0f, 0.0f} },
             { Vector2{0.5f, 0.5f},   Vector3{0.0f, 0.0f, 0.0f} },
@@ -56,7 +46,7 @@ namespace nc::graphics
         m_indexBuffer = std::make_unique<IndexBuffer>(*m_base, *m_commands, indices);
 
         // Write the render pass to the command buffer.
-        m_commands->RecordRenderCommand(*m_base, *m_renderPass, *m_frameBuffers, *m_pipeline, *m_vertexBuffer, *m_indexBuffer);
+        m_commands->RecordRenderCommand(*m_pipeline, *m_swapchain, *m_vertexBuffer, *m_indexBuffer);
     }
 
     Graphics2::~Graphics2() = default;
@@ -67,23 +57,12 @@ namespace nc::graphics
         WaitIdle();
 
         // Destroy all resources used by the swapchain
-        m_commands.reset();
-        m_frameBuffers.reset();
-        m_pipeline.reset(); // @todo: Set up dynamic state in pipeline for scissor rect and viewport so entire pipeline is not recreated on resize.
-        m_renderPass.reset();
-        m_base->CleanupSwapChain();
+        m_base->FreeCommandBuffers();
+        m_swapchain->CleanupSwapChain();
 
         // Recreate swapchain and resources
-        m_base->CreateSwapChain(dimensions);
-        m_renderPass = std::make_unique<RenderPass>(*m_base);
-        m_pipeline = std::make_unique<GraphicsPipeline>(*m_base, *m_renderPass);
-        m_frameBuffers = std::make_unique<FrameBuffers>(*m_base, *m_renderPass);
-        m_commands = std::make_unique<Commands>(*m_base,
-            m_base->GetSemaphores(SemaphoreType::RenderReady),
-            m_base->GetSemaphores(SemaphoreType::PresentReady),
-            m_base->GetFences(FenceType::FramesInFlight),
-            m_base->GetFences(FenceType::ImagesInFlight));
-        m_commands->RecordRenderCommand(*m_base, *m_renderPass, *m_frameBuffers, *m_pipeline, *m_vertexBuffer, *m_indexBuffer);
+        m_swapchain->RecreateSwapchain(dimensions);
+        m_commands->RecordRenderCommand(*m_pipeline, *m_swapchain, *m_vertexBuffer, *m_indexBuffer);
     }
 
     DirectX::FXMMATRIX Graphics2::GetViewMatrix() const noexcept
@@ -142,10 +121,10 @@ namespace nc::graphics
 
     bool Graphics2::GetNextImageIndex(uint32_t& imageIndex)
     {
-        m_base->WaitForFrameFence();
+        m_swapchain->WaitForFrameFence();
 
         bool isSwapChainValid = true;
-        imageIndex = m_base->GetNextRenderReadyImageIndex(isSwapChainValid);
+        imageIndex = m_swapchain->GetNextRenderReadyImageIndex(isSwapChainValid);
         if (!isSwapChainValid)
         {
             RecreateSwapChain(m_dimensions);
@@ -156,16 +135,16 @@ namespace nc::graphics
 
     void Graphics2::RenderToImage(uint32_t imageIndex)
     {
-        m_base->WaitForImageFence(imageIndex);
-        m_base->SyncImageAndFrameFence(imageIndex);
-        m_base->ResetFrameFence();
-        m_commands->SubmitRenderCommand(*m_base, imageIndex);
+        m_swapchain->WaitForImageFence(imageIndex);
+        m_swapchain->SyncImageAndFrameFence(imageIndex);
+        m_swapchain->ResetFrameFence();
+        m_commands->SubmitRenderCommand(imageIndex);
     }
 
     bool Graphics2::PresentImage(uint32_t imageIndex)
     {
         bool isSwapChainValid = true;
-        m_base->Present(imageIndex, isSwapChainValid);
+        m_swapchain->Present(imageIndex, isSwapChainValid);
 
         if (!isSwapChainValid)
         {
@@ -198,6 +177,6 @@ namespace nc::graphics
     void Graphics2::FrameEnd()
     {
         // Used to coordinate semaphores and fences because we have multiple concurrent frames being rendered asynchronously
-        m_base->IncrementFrameIndex();
+        m_swapchain->IncrementFrameIndex();
     }
 }
