+++ nc/source/graphics/vulkan/Commands.h
@@ -6,27 +6,27 @@
 
 namespace nc::graphics::vulkan
 {
-    class Base; class RenderPass; class FrameBuffers; class GraphicsPipeline; class VertexBuffer; class IndexBuffer;
+    class Base; class FrameBuffers; class GraphicsPipeline; class VertexBuffer; class IndexBuffer; class Swapchain;
 
     class Commands
     {
         public:
-            Commands(const vulkan::Base& base, 
-                     const std::vector<vk::Semaphore>& renderReadySemaphores, 
-                     const std::vector<vk::Semaphore>& presentReadySemaphores, 
-                     const std::vector<vk::Fence>& framesInFlightFences, 
-                     const std::vector<vk::Fence>& imagesInFlightFences);
+            Commands(const vulkan::Base& base, const vulkan::Swapchain& swapchain);
 
-            void RecordRenderCommand(const vulkan::Base& base, const vulkan::RenderPass& renderPass, const FrameBuffers& frameBuffers, const vulkan::GraphicsPipeline& pipeline, const vulkan::VertexBuffer& vertexBuffer, const vulkan::IndexBuffer& indexBuffer);
-            void SubmitRenderCommand(const vulkan::Base& base, uint32_t imageIndex);
-
-            void SubmitCopyCommandImmediate(const vulkan::Base& base, const vk::Buffer& sourceBuffer, const vk::Buffer& destinationBuffer, const vk::DeviceSize size);
+            void RecordRenderCommand(const vulkan::GraphicsPipeline& pipeline, const vulkan::Swapchain& swapchain, const vulkan::VertexBuffer& vertexBuffer, const vulkan::IndexBuffer& indexBuffer);
+            void SubmitRenderCommand(uint32_t imageIndex);
+            void SubmitCopyCommandImmediate(const vk::Buffer& sourceBuffer, const vk::Buffer& destinationBuffer, const vk::DeviceSize size);
 
         private:
-            std::vector<vk::CommandBuffer> m_commandBuffers;
+            // External members
+            const Base& m_base;
+            const Swapchain& m_swapchain;
             const std::vector<vk::Semaphore>& m_renderReadySemaphores;
             const std::vector<vk::Semaphore>& m_presentReadySemaphores;
             const std::vector<vk::Fence>& m_framesInFlightFences;
             const std::vector<vk::Fence>& m_imagesInFlightFences;
+
+            // Internal members
+            std::vector<vk::CommandBuffer> m_commandBuffers;
     };
 }
