--- nc/source/graphics/vulkan/RenderPass.h
@@ -1,17 +1,0 @@
-#pragma once
-
-#include "vulkan/vulkan.hpp"
-
-namespace nc::graphics::vulkan
-{
-    class Base;
-
-    class RenderPass
-    {
-        public:
-            RenderPass(const vulkan::Base& base);
-            const vk::RenderPass& GetRenderPass() const noexcept;
-        private:
-            vk::UniqueRenderPass m_renderPass;
-    };
-}
