+++ nc/source/graphics/vulkan/Resources/IndexBuffer.h
@@ -14,7 +14,8 @@ namespace nc::graphics::vulkan
     class IndexBuffer
     {
         public:
-            IndexBuffer(Base& device, Commands& commands, std::vector<uint32_t> indices);
+
+            IndexBuffer(Base& base, Commands& commands, std::vector<uint32_t> indices);
             void Bind();
             uint32_t GetId() const;
             const vk::Buffer& GetBuffer() const;
@@ -22,6 +23,7 @@ namespace nc::graphics::vulkan
             const std::vector<uint32_t>& GetIndices() const;
 
         private:
+
             vulkan::Base& m_base;
             vulkan::Commands& m_commands;
             uint32_t m_id;
